pyproject.toml
[build-system]
requires = ["setuptools", "numpy"]
build-backend = "setuptools.build_meta"


[project]
name = "laueimproc"
dynamic = ["version"]
requires-python = ">=3.10"
dependencies = [
  "click",
  "cloudpickle",
  "numpy",
  "opencv-contrib-python-headless",
  "pillow",
  "psutil",
  "pyyaml",
  "torch",
  "tqdm",
]
authors = [
  {name = "Robin RICHARD (robinechuca)", email = "serveurpython.oz@gmail.com"},
]
maintainers = [
  {name = "Robin RICHARD (robinechuca)", email = "serveurpython.oz@gmail.com"},
]
readme = "README.rst"
license = {file = "LICENSE"}
keywords = [
  "laue",
  "multigrain",
]
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Environment :: GPU",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Natural Language :: English",
  "Natural Language :: French",
  "Programming Language :: Python :: 3",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Scientific/Engineering :: torch.Tensor Processing",
  "Topic :: Scientific/Engineering :: Information Analysis",
  "Topic :: Scientific/Engineering :: Physics",
]

[project.optional-dependencies]
all = [
  "manim",
  "matplotlib",
  "pdoc3",
  "pydocstyle<=6.1.1",  # 6.1.2 bug on pylama 8.4.1
  "pylama[all]",
  "pylama[toml]",
  "pytest",
  "recommonmark",
  "sphinx",
  "sphinx-copybutton",
  "sphinx-inline-tabs",
  "sphinx-rtd-dark-mode",
  "sphinx-rtd-theme",
  "sphinxcontrib-video",
  "sphinxext-opengraph",
]

[project.scripts]
laueimproc = "laueimproc.__main__:main"
laueimproc-convert = "laueimproc.io.__main__:main"
laueimproc-doc = "laueimproc.__main__:main_doc"
laueimproc-test = "laueimproc.testing.__main__:main"

[project.urls]
Repository = "https://github.com/BM32ESRF/laueimproc"

[tool.setuptools.packages.find]
where = [""]
exclude = ["doc", "notebook"]

[tool.setuptools.package-data]
"*" = ["*.jp2"]

[tool.setuptools]
py-modules = ["c_build"]

[tool.setuptools.cmdclass]
build_py = "c_build.Build"

[tool.setuptools.dynamic]
version = {attr = "laueimproc.__version__"}

[tool.pytest.ini_options]
filterwarnings = [
  "ignore:Module already imported:pytest.PytestAssertRewriteWarning"
]
markers = [
  "slow: marks tests as slow",
]

[tool.pylama]
sort = "E,W,D"
skip = "./doc, ./dist*"
linters = "mccabe,pycodestyle,pydocstyle,pyflakes,mypy,pylint"
# E226 false detection pycodestyle 2.11.1
# D212 and D213 are mutualy exclusive, PEP 257 is in favor of D212
# D211 and D203 are mutualy exclusive, PEP 257 is in favor of D211
# D107 is in contradiction with the numpy doc convention
# D413 disable in the google, numpy, and pep257 convention
ignore = "E226,D213,D203,D107,D413"
max_line_length = 100

[tool.mypy]
cache_dir = "~/.cache/mypy_cache"
ignore_missing_imports = true
show_error_codes = true

[tool.pylint]
# E0611 failed to import c compiled modules
# C0103 method name not conform because qt impose the name
# E1101 'no member' failed in all c extensions
# E1102 'not callable' false detection in pylint 3.1.0
# E1133 'no iterable' false positive in pylint 3.1.0
# E1135 'doesn't support membership test' false positive in pylint 3.1.0
# C0415 'import outside toplevel' solve cyclic import and allow optional modules
disable = "E0611,C0103,E1101,E1102,E1133,E1135,C0415"
extension-pkg-allow-list = [
  "laueimproc.geometry.c_metric",
  "laueimproc.gmm.c_fit",
  "laueimproc.gmm.c_gmm",
  "laueimproc.improc.c_find_bboxes",
  "laueimproc.improc.spot.c_basic",
  "laueimproc.improc.spot.c_extrema",
  "laueimproc.improc.spot.c_pca",
  "laueimproc.ml.c_dist",
  "laueimproc.opti.c_rois",
]
# enable = "all"
